Refactor parser test => reader test, split out testExpectations
